+++ Source/MarshmallowCocoa/NSString+Marshmallows.m
@@ -19,7 +19,7 @@
     OSStatus error = anOSStatus;
     
     // see if it appears to be a 4-char-code
-    *(UInt32 *)(str + 1) = CFSwapInt32HostToBig(error);
+    *(UInt32 *)(str + 1) = CFSwapInt32HostToBig((u_int32_t)error);
     if (isprint(str[1]) && isprint(str[2]) && isprint(str[3]) && isprint(str[4])) {
         str[0] = str[5] = '\'';
         str[6] = '\0';
